# src/main/java/gift/category/dto/CategoryResponse.java
package gift.category.dto;

import gift.category.entity.Category;

public class CategoryResponse {
    private Long id;
    private String name;
    private String color;
    private String imageUrl;
    private String description;

    public CategoryResponse(Category category) {
        this.id = category.getId();
        this.name = category.getName();
        this.color = category.getColor();
        this.imageUrl = category.getImageUrl();
        this.description = category.getDescription();
    }

    public CategoryResponse(Long id, String name, String color, String imageUrl,
        String description) {
        this.id = id;
        this.name = name;
        this.color = color;
        this.imageUrl = imageUrl;
        this.description = description;
    }

    public Long getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public String getColor() {
        return color;
    }

    public String getImageUrl() {
        return imageUrl;
    }

    public String getDescription() {
        return description;
    }
}